Added a new output type for simple flat files.
This output type writes one JSON per line in flat files and rotates
these files similarly with the way syslog does. This can be useful for
integrating Packetbeat with other tools, see #27.

Sharing topology between agents is not supported when using only this output
type, but it can be used together with the REDIS one to also provide this
functionality.