The sniffer is now a real input plugin.
The sniffer plugin still needs to be a bit special because
it has a few command line flags, because the socket needs
to be opened before dropping privileges and because it will
call the protocol plugins directly instead of writing in
the channel.

Still, much better than having the logic in the main method.

This also means the sniffing logic is moved out of the main
goroutine. I don't expect any side-effects from that, but
needs testing.