Insert in ES a document for each process
There are two types of documents that are inserted in ES:
1. index=topbeat-*, type=system for system wide statistics (memory, swqp, cpu times, system load)
2. index=topbeat-*, type=proc for per process statistics (memory, cpu times, ppid, state, process name)
in order to be able to easily build per process statistics in Kibana, like
the graphs describing the evolution of CPU usage for a certain process.

By default, all the processes that are running on a server are exported to ES.
You can configure a subset of processes by using regular expressions.

The default configuration is:
input:
 period: 1 // every second statistics
 procs: [".*"] // all processes are exported to ES