Generate a BPF filter that can skip VLAN tags.
Generates BPF filters like this: "port 80 and (vlan and port 80)"
which is needed for capturing traffic that has 802.1Q tags.

This is done only optionally because, for example, if you try to
set this for the loopback interface on OS X, you will get an error.

Also adds the option to overwrite the BPF filter entirely from
the configuration file. This can be useful for excluding entire
hosts or networks from the packet capture at the lowest level.