Linux compile on Debian 6.
We used to compile on a recent ubuntu version, which is needed for
cross-compiling to Windows and OS X. This poses the problem that it links
against a relatively recent version of libc, so the resulting binary cannot work
on older linux distributions. To work around this issue we tried to compile
statically linked binaries. Unfortunately, statically linking against libc
breaks in part the name resolution, in particular resolving the names from
/etc/hosts.

This pull requests makes 2 Docker images: one based on ubuntu for cross
compiling to os x and windows and one based on debian 6 for compiling for Linux
(both 32 and 64 bits versions). Libpcap continues to be statically linked for
Linux.

Because the Dockerfiles have drifted quite a bit from the versions from XGO, I
moved them to this repository.

This closes elastic/packetbeat#225.