Fix negative response times for MongoDB.
The message timestamp could have was accidentally reused from an
old message in the same stream. For example, in this sequence of
messages in the same stream:

* `REPLY1` (with no matching request)
* `REQUEST2`
* `REPLY2`

packetbeat was correctly matching the REQUEST2 with REPLY2 but was using
the timestamp of REPLY1 for computing the response time. The fix is to
simply through away the message and wait for it to be created with the
new message received.

Should fix most of the issues from #216.

This also removes parseOffset and bytesReceived from the stream struct as
they were not used.