Appveyor configuration
Runs unit and integration tests on the Appveyor service.

The approach taken installs Go and python and then run the equivalent commands
that the Makefile would run for building and testing. This means a bit of code
duplication but using make on Windows is hard enough and maintaining a
cross-platform Makefile would be a PITA.