Use Travis to build packages and upload them to S3
To make this more convenient for travis, the image building is
no longer a Makefile dependency. Instead, the following targets
are added:

* make images - build images from docker file
* make push-images - push images to the docker hub
* make pull-images - pull images from the docker hub

The push requires the credentials for the tudorg account, but Travis
has them, so you can (ab)use it to rebuild the images if we need to
change something in the Dockerfiles.