Single Vagrantfile for all Beats
Adding a single Vagrantfile to the root of the project that can be used by all beats for development. The Windows box has been updated. In addition to having golang and python like the previous one, this box has mingw64, winpcap, SSH (for vagrant ssh), and a desktop shortcut that launches a command prompt in the beats directory.