Unique version constant for the official Beats
This makes libbeat have a default for the version string. The beats
can still overwrite this, but the Beats from the main repo won't, because
their versions are in sync.

Main motivation is to make this friendlier to automatic modifying of the
version. That is also why the constant is split in its own file.