Refactor prospector to split up different types
* Simplify prospector and introduce prospectorer interface. Like this every prospector type runs independent.
* Change logic on how prospector is initalised and started. See below.

Changing logic on how prospectors are setup

Before prospectors were set up the ones finished could already start to send data. This made it necessary to start the Registrar after prospector startup to make sure not states were already written.
Now it is waited until the setup of all prospectors is completed. The reasoning behind this is that this only happens on startup and no prospector should take a long time to initialise. Afterwards all prospectors can start to send data in parallel.

This simplifies the code and reduces the dependencies.

In case we use one registrar/spooler per prospector in the future (for example for performance improvements), the init phase could also be decoupled