Metricbeat config refactoring
This change to the configuration file makes it possible to configure each module and metricset multiple times. This allows to run each metricset of a module independent of the module with different configuration. In addition it allows to run a metricset with different configuration like period for different hosts.

An example use case can be, that the basic apache metricset is run every second for localhost, but the extended apache metricset for localhost is only run every 10 seconds. This also allows to add different filters and fields if needed.

Based on this config change, metricbeat was refactored to decouple modules and metricsets.