Add full metricset and docs for redis
Currently all data that is reported by redis is sent as an event.

In a second step the notion of a level like minimal, basic, full should be introduced that can be configured so not all data is sent.
It is always possible to use drop_fileds to minimize the number of fields sent. Fields are nested to make it easier for example to remove all server information.