Decouple Data Fetch part in Topbeat
The data fetching and processing part was completely decoupled from Topbeat. This allows to reuse the functionality in metricbeat or other beats.

For the moment the variable functions stayed the same even though they could be simplified in a second step. The goal is to keep the changes as low as possible.