Create a MetricSet per host. (#1586)
This change ensures a single MetricSet instance is used per host. Creating a MetricSet per host simplifies MetricSet implementations because the MetricSet does not need to concern itself with multiple hosts. Previously a single MetricSet instance was reused for fetching data from multiple hosts. It's Fetch() method was invoked by various goroutines so any mutable state had to be made concurrency-safe which added a source of complexity. Now since a MetricSet is used by only one host, it's Fetch() method is only invoked by a single goroutine so no synchronization is required to protect data members.

Added validation to ensure that the same host was not defined more than once in the hosts configuration.

Fixed an issue where if multiple instances of the same module type were used, that only the expvar metrics from the last one would be exported.