Set version to 5.0.0-alpha3 (#1718)
* Set version to 5.0.0-alpha3

To be closer to what we plan for the unified release process, I'm setting the
version in advance. The artifacts will be named 5.0.0-alpha3-SNAPSHOT during
testing but the version in code is 5.0.0-alpha3. At the time of the release, we
rename the artifacts to remove SNAPSHOT, without having to rebuild.

* Fix set_version script to remove leading empty line