Add support to import only the index-pattern or only the dashboards
Add -only option to be able to import only dashboards or index-pattern. By default both, dashboards and index-pattern
are imported.

Fix script to be able to import dashboards from a local directory. For this, -dir option is used only to import from a
local directory, and it's not used in addition to the -url or -file options. If the dashboards archive contains dashboards
for multiple Beats, then you can specify the Beat name under the -beat option.