Update docker client to most recent version (#3398)
* Update docker client to most recent version

The docker client now depends on the docker repository which means we have to add the docker repo. This adds quite a few dependencies. To reduce the number of dependencies, the docker repository was modified manually and all not required parts were removed. On the client side some code parts were removed which are not needed for our read only use case.

Long term goal is to either use the internal docker client or use our own implementation of a HTTP client that fetches data from the socket or rest API and uses our internal schema.

* remove some more files

add removal of .tar files to cleanup script

* removing more files to make tests happy