Add process summary metrics (#4231)
As we're enabling `include_top_n` by default, a few visualizations from the
Metricbeat-processes list were no longer correct (as they aggregate only a
sample of the data).

This adds a new `process_summary` metricset that adds these metrics. The fields
are namespaced under `process.summary`.

This PR adds summary metrics for the total number of processes and their state, as an
extra document created by the `system.process` metricset.