Enable specifying pip install options (#5693)
If a binary wheel is available, pip will prefer that to building from
source. Yet some packagers prefer to build everything from source. For
this, the ability to specify the --no-binary flag is needed for
pip. This commit enables this possibility by enabling the
PIP_INSTALL_PARAMS environment variable to be specified so that
packagers can specify PIP_INSTALL_PARAMS="--no-binary :all:".