Make Metricbeat framework more configurable (#5835)
* Make Metricbeat framework more configurable

This makes the Metricbeat framework more flexible by allowing developers to specify options when instantiating Metricbeat. This adds an option for controlling the event format and whether or not the traditional `metricset` info object is added to each event that contains the metricset name, module name, host, and round-trip-time (rtt).

Additionally this adds an Event struct that MetricSets can use to deliver their data. The Event struct has fields for all data that was previously sent in one `MapStr` by using dedicated key names like `_namespace`, `_rtt`, `_module`, and `@timestamp`. This makes the MetricSet code a bit more clear.

Because of the introduction of the `mb.Event` struct I'm deprecating the `mb.Reporter` interface that contained methods for reporting data using a `MapStr`. The `mb.Reporter` interface defines the methods `Event(MapStr)`, `Error(error)`, and `ErrorWith(error, MapStr)`. This was added recently to enable push metricsets so it isn't widely used yet. After a short deprecation period this interface will be removed.  This interface is replaced with `mb.ReporterV2` that contains `Event(mb.Event)` and `Error(error)`.

* Disable logging during module testing

Logging statements were interleaved with expected test output.
Report an error when error.message exists.

* Add WithNamespace as MetricSet option

`WithNamespace` can be used when registering a MetricSet to specify the namespace under which the MetricSet's data should be written. This only needs to be used when the namespace differs from [module].[metricset]. The value contains the fully qualified namespace so for the elasticsearch node_stats it would be `WithNamespace("elasticsearch.node.stats")`. This should only be used with MetricSets that use the `ReporterV2` interface.

* Add PushMetricSetV2 testing utils