Remove 2x environments (#5967)
As by now 6.x releases are out, I removed the 2.x environments we used for manual testing. They are not needed anymore and were not used for quite some time.

This also removed some of the Dockerfile which now makes it possible to rely on only 1 Docker file for the different versions of the pieces in the Elastic Stack.