Change format of exported fields content (#6861)
This PR updates the generate_fields_docs.py to make the layout of the reference content easier to read. It uses definition lists instead of sections to format the options. This makes the content easier to scan, and it makes the "On this page" section more useful.

Here's an example. I'm still not happy with repeated section names, but this PR at least makes the content easier to read.

![image](https://user-images.githubusercontent.com/14206422/38521498-b7f9c0fe-3bfa-11e8-8f9e-c6cb060b4d53.png)