Port fields.yml collector to Golang (#6911)
Existing Python and shell code is ported to Golang. The previous Beat specific field collection is generalized and moved to the `Makefile` of `libbeat`. A new variable is added to makefiles: `FIELDS_FILE_PATH`. This specifies where `fields.yml` are.

Packetbeat's `fields_base.yml` is renamed to `fields.common.yml` to be the same as in other Beats