Update export_dashboard script to export full dashboard json doc (#7241)
In https://github.com/elastic/beats/pull/7239 support to export a dashboard is a added to each Beat. This is expected to be used by users. The export_dashboard script from the Beats repository is expected by the Devs and contributors which want to add new dashboards.

In https://github.com/elastic/beats/pull/7224 the dashboards are modified to be stored with decoded json objects for better versioning. This change modifies the export dashboard script so it generates the same decoded output so no additional conversion is needed.