Add generation of generated files for Filebeat module tests (#7430)
Filebeat module tests allow to provide an `*-expected.json` file to compare the ingested data with the expected data. So far these files had to be generated manually by exporting the data from Elasticsearch. To make the generation easier now the environment variable `GENERATE=1` can be used and the existing file will be overwritten.

As the files are now generated, the structure of the files changed a little bit:

* fields like source path, event.created, beat.hostname are removed from the events as they change between tests.
* Objects are flattened for easier comparison
* The header response part with `_id` etc. was removed as it's not needed.

All expected.json files are overwritten in this PR to follow the new schema.

An example command to overwrite the generated files for Kibana looks as following:
```
INTEGRATION_TESTS=1 TESTING_FILEBEAT_MODULES=kibana GENERATE=1 nosetests tests/system/test_modules.py
```

Others:

* Fix naming of Kibana expected file.
* Remove extended check environment variable as all modules go now through the extended check.
* Skip @timestamp comparison for filesets without timestamp in log line.