Add kubernetes to services which are kept running (#7458)
* Add kubernetes to services which are kept running

Our metricbeat tests on Jenkins are flaky because of Kubernetes. My current assumption is that there are some overlaps when starting / stopping kuberenetes. Reason for this is that we do not see the same issue on Travis. The different between Travis and Jenkis is probably the number of processes which run in parallel. So either we could reduce the number of processes for the tests to 1 or add kubernetes to the services that keep running. I decided to keep it running for now to see what the outcome is.