Add MacOS installer to packaging
This adds the creation of a .dmg files when building packages on MacOS (the
build host must be darwin with XCode installed). The .dmg file contains an
installer (.pkg) and a rudimentary Uninstall.app.  When code signing is enabled
every part of the .dmg is signed. This is the structure of the package.

- beat.dmg
  - beat.pkg
    - internal-beat.pkg
    - BeatPrefPane.pkg
  - Uninstall.app

The following environment variables control the code signing behavior.

- APPLE_SIGNING_ENABLED - Must be set to true to enable signing. Defaults to
  false.
- APPLE_SIGNING_IDENTITY_INSTALLER - filter for selecting the signing identity
  for installers. It's effectively used as
  `security find-identity -v | grep $APPLE_SIGNING_IDENTITY_INSTALLER` to
  select a single certificate.
- APPLE_SIGNING_IDENTITY_APP - filter for selecting the signing identity
  for apps.

Co-authored-by:  Adrian Serrano <adrisr83@gmail.com>