Allow to specify auditd rules in separate files (#7331)
* Auditbeat: Allow to specify auditd rules in separate files (#7113)

Added a new configuration field, `audit_rule_files` to the auditd
module. It allows to specify a set of files that contain audit rules.

- module: auditd
  audit_rule_files: [ '${path.config}/auditd.rules.d/*.conf' ]

Rules will be loaded into the kernel in the following order:
1. Rules specified in the `audit_rules` field at auditbeat.yml.
2. Rules from files in the `audit_rule_files`, in the order they are
   listed.
3. If a path contains a glob, the resulting files are loaded in
   lexicographical order.

Closes #7113

* Install sample rules in /etc/auditbeat/audit.rules.d

Updated the build scripts to install sample audit rules in auditbeats
configuration directory. Different rules are used for 32 and 64 bit
platforms.